add report load() and save() with pickle
store and retrieve reports from the fs using pickle!
Report().save(".db.obj") writes Report() to the fs as .db.obj
which can be loaded using the pickle module, or using
Report().load(".db.obj") which uses a hack to surgically update
self with the pickle loaded object.

testing for such with the new pytest framework